--- frontend/src/components/User/Onboarding/Onboarding.css
@@ -1,194 +1,0 @@
-.container{
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    align-items: center;
-    height: 100vh;
-    width: 900px; 
-    margin: 0 auto;
-}
-
-.container{
-    margin-top: 10px;
-}
-.card{
-    border: none;
-}
-
-.steps {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    position: relative;
-}
-.step-button {
-    
-    width: 30px;
-    height: 30px;
-    border-radius: 50%;
-    background-color: white;
-    color: var(--black);
-    border: 1px solid var(--prm-color);
-    transition: .4s;
-}
-.step-button[aria-expanded="true"] {
-    width: 50px;
-    height: 50px;
-    background-color: var(--prm-color);
-    color: #fff;
-}
-.card-body{
-    background-color: var(--bgc-color);
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    animation: scale-up-center 0.4s cubic-bezier(0.390, 0.575, 0.565, 1.000) both;
-}
-.text-card{
-    text-align: center;
-}
-.text-card > h2 > span{
-    font-weight: 600;
-    color: var(--prm-color);
-}
-.text-card > p{
-    font-family: 'Montserrat';
-    font-style: normal;
-    font-weight: 400;
-    font-size: 24px;
-    line-height: 32px;
-}
-.text-card > p > span{
-    font-weight: 500;
-}
-.text-card > ul{
-    list-style-position: inside;
-    font-style: normal;
-    font-weight: 400;
-    font-size: 20px;
-    line-height: 32px;
-}
-.logo{
-    text-align: center;
-    background-color: var(--bgc-color);
-}
-.done {
-    background-color: var(--prm-color);
-    color: white;
-    border: 1px solid var(--black);
-}
-.step-item {
-    z-index: 10;
-    text-align: center;
-}
-#progress {
-  -webkit-appearance:none;
-    position: absolute;
-    width: 97%;
-    z-index: 5;
-    height: 10px;
-    margin-left: 20px;
-    margin-bottom: 18px;
-    margin-top: 21px;
-}
-
-.btn-card-1{
-    display: flex;
-    flex-direction: column;
-    background-color: var(--bgc-color);
-}
-.btn-card-1 > button{
-    margin: auto;
-    width: 370px;
-    height: 40px;
-    margin-bottom: 10px;
-    border-radius: 4px;
-}
-.btn-1{
-    background-color: #FE662E;
-    color: white;
-    border: 1px solid var(--black);
-    box-shadow: 0 5px 5px rgba(0, 0, 0, 0.521);
-}
-.btn-2{
-    color: var(--azul-color);
-    border: 1px solid var(--azul2-color);
-}
-.fa-arrow-right-long{
-    margin-left: 15px;
-
-}
-.cardtwo{
-    display: flex;
-    align-items: flex-start;
-    margin-top: 80px;
-}
-#img-design, #img-develop{
-    display: none;
-}
-.quest{
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    text-align: center;
-}
-.quest > div{
-    margin:30px 0;
-}
-.pagDesign, .pagDevelop{
-    display: none;
-    flex-wrap: wrap;
-    justify-content: center;
-    width: 90%;
-}
-
-/* para personalizar a barra de progresso */
-#progress::-webkit-progress-value {
-    background-color: var(--prm-color);
-    transition: .5s ease;
-}
-#progress::-webkit-progress-bar {
-    background-color: var(--prm-gray);
-
-}
-
-@keyframes scale-up-center {
-    0% {
-      transform: scale(0.5);
-    }
-    100% {
-      transform: scale(1);
-    }
-  }
-
-.ck-button {
-    margin:5px;
-    background: rgba(255, 255, 255, 1e-06);
-    border-radius:20px;
-    border: 1px solid #474747;;
-    margin-right: 20px;
-    overflow:auto;
-    float:left;
-}
-
-.ck-button label span {
-    text-align:center;
-    padding: 7px 13px;
-    display:block;
-    border-radius:20px;
-    
-}
-
-.ck-button label input {
-    visibility: hidden;
-    position:absolute;
-    top:-20px;
-    
-}
-input:checked +span{
-    background-color: #FE662E;
-    color:#fff;
-}
-.ck-button label:hover   {
-    background-color:#F8F9FA;
-}
