+++ onboarding.html
@@ -36,32 +36,6 @@
                 </div>
                 
                 <div class="card">
-                
-                    <div id="collapseOne" class="collapse show" aria-labelledby="headingOne"
-                        data-bs-parent="#accordionExample">
-                        <div class="card-body">
-                        <div>
-                            <img src="./frontend/src/assets/svg/teampic.svg" alt="">
-                        </div>
-                        <div class="text-card">
-                            <h2>Olá <span>Sangue Laranja</span>, que bom ter você aqui!</h2>
-                            <p>Você está prestes a se conectar à <span>ROBSON</span>: Rede Otimizada de Busca e Solução “Orange Networking”</p>
-                            <ul>
-                                <li>Pessoas da FCamara com diversas skills prontas pra te ajudar ou precisando de sua ajuda.</li>
-                                <li>Grupos de estudos (crowdlearning) para compartilhar suas experiências e aprender coletivamente.</li>
-                                <li>Materiais indicados pelos usuários da ROBSON.
-                                </li>
-                            </ul>
-                        </div>
-                        <div class="btn-card-1">
-                            <button type="button" class="btn-1" id="btn-1" data-bs-toggle="collapse"
-                            data-bs-target="#collapseTwo">Vamos lá!<i class="fa-solid fa-arrow-right-long"></i></button>
-                            <button class="btn-2">Não quero me conectar</button>
-                        </div>
-                        </div>
-                    </div>
-                </div>
-                <div class="card">
                     <div id="collapseTwo" class="collapse" aria-labelledby="headingTwo" data-bs-parent="#accordionExample">
                         <div class="card-body">
                             <div class="cardtwo">
